Move test setup into mock, rename tests to use buy/sell convention, minor test corrections
